requirements.txt
python-dotenv==1.0.0
flask==3.0.0 
openai==0.28.0
google-cloud-aiplatform==1.39.0
vertexai==0.0.1
pydub==0.25.1
PyMuPDF==1.23.12
langchain==0.0.353
langchain-core==0.1.4
pinecone-client==2.2.4
pypdf==3.17.4
tiktoken==0.5.2
googletrans==3.1.0a0